wc - Moved opencga-knockout-folder to deprecated
