+++ src/webcomponents/variant/analysis/deprecated/opencga-knockout-analysis/knockout-individual-view.js
@@ -15,13 +15,13 @@
  */
 
 import {LitElement, html} from "lit";
-import UtilsNew from "../../../../core/utils-new.js";
-import CatalogGridFormatter from "../../../commons/catalog-grid-formatter.js";
+import UtilsNew from "../../../../../core/utils-new.js";
+import CatalogGridFormatter from "../../../../commons/catalog-grid-formatter.js";
 import AnalysisRegistry from "../analysis-registry.js";
-import GridCommons from "../../../commons/grid-commons.js";
+import GridCommons from "../../../../commons/grid-commons.js";
 import knockoutDataIndividuals from "../test/knockout.20201103172343.kFIvpr.individuals.js";
 import "./knockout-individual-variants.js";
-import "../../../family/family-view.js";
+import "../../../../family/family-view.js";
 
 export default class KnockoutIndividualView extends LitElement {
