+++ src/webcomponents/variant/analysis/deprecated/opencga-knockout-analysis/knockout-variant-individual.js
@@ -15,9 +15,9 @@
  */
 
 import {LitElement, html} from "lit";
-import UtilsNew from "../../../../core/utils-new.js";
-import GridCommons from "../../../commons/grid-commons.js";
-import "./../../../commons/view/detail-tabs.js";
+import UtilsNew from "../../../../../core/utils-new.js";
+import GridCommons from "../../../../commons/grid-commons.js";
+import "../../../../commons/view/detail-tabs.js";
 import knockoutDataIndividuals from "../test/knockout.20201103172343.kFIvpr.individuals.js";
 
 
@@ -54,8 +54,7 @@ export default class KnockoutVariantIndividual extends LitElement {
     }
 
     updated(changedProperties) {
-        if (changedProperties.has("opencgaSession")) {
-        }
+        // if (changedProperties.has("opencgaSession")) {}
 
         if (changedProperties.has("variant")) {
             this.prepareData();
@@ -86,7 +85,7 @@ export default class KnockoutVariantIndividual extends LitElement {
             uniqueId: "id",
             pagination: true,
             paginationVAlign: "both",
-            //formatShowingRows: this.gridCommons.formatShowingRows,
+            // formatShowingRows: this.gridCommons.formatShowingRows,
             gridContext: this,
             formatLoadingMessage: () => "<div><loading-spinner></loading-spinner></div>",
             onClickRow: (row, selectedElement, field) => {
