+++ src/webcomponents/variant/analysis/deprecated/opencga-knockout-analysis/knockout-variant-view.js
@@ -15,17 +15,17 @@
  */
 
 import {LitElement, html} from "lit";
-import UtilsNew from "../../../../core/utils-new.js";
-import CatalogGridFormatter from "../../../commons/catalog-grid-formatter.js";
+import UtilsNew from "../../../../../core/utils-new.js";
+import CatalogGridFormatter from "../../../../commons/catalog-grid-formatter.js";
 import AnalysisRegistry from "../analysis-registry.js";
-import GridCommons from "../../../commons/grid-commons.js";
+import GridCommons from "../../../../commons/grid-commons.js";
 import knockoutData from "../test/knockout.20201029141213.SChLEA.js";
 import "./knockout-individual-variants.js";
 import "./knockout-variant-allele-pairs.js";
 import "./knockout-variant-individual.js";
-import "../../../family/family-view.js";
-import "../../annotation/cellbase-population-frequency-grid.js";
-import "../../annotation/variant-annotation-clinical-view.js";
+import "../../../../family/family-view.js";
+import "../../../annotation/cellbase-population-frequency-grid.js";
+import "../../../annotation/variant-annotation-clinical-view.js";
 
 
 export default class KnockoutVariantView extends LitElement {
